requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

pyirsdk==1.3.3
